image2pdf/requirements.txt
PyPDF2 
Pillow
